% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.R
\name{AA}
\alias{AA}
\title{transfer nucleic acid sequence to amino acids}
\usage{
AA(x)
}
\arguments{
\item{x}{a text string containing a nucleic acid sequence}
}
\value{
a text string that represents the translated amino acid sequence
}
\description{
transfer nucleic acid sequence to amino acids
}
\examples{
x<-"ATGtacaca"
AA(x)
}
